// src/types/generated/Hint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Pos } from "./Pos";
import type { StrategyType } from "./StrategyType";

export type Hint = { strategy: StrategyType, target: Pos, value: number, related_cells: Array<Pos>, eliminated: Array<number>, message_cn: string, };